+++ flutter_service_bloc/example/lib/open_library/constant/constant.dart
@@ -1,0 +1,1 @@
+part 'uri_constant.dart';