+++ flutter_service_bloc/example/lib/open_library/constant/uri_constant.dart
@@ -1,0 +1,6 @@
+part of 'constant.dart';
+
+const String apiDomain = 'openlibrary.org';
+
+const String searchAuthorPath = '/search/authors.json';
+const String authorDetailPath = '/authors';