+++ flutter_service_bloc/lib/src/service_bloc_builder.dart
@@ -13,59 +13,17 @@ class ServiceBlocBuilder<
     Bloc extends ServiceBloc<ServiceRequestedEvent, Response>,
     ServiceRequestedEvent extends ServiceRequested,
     Response> extends BlocBuilder<Bloc, ServiceState> {
-  /// A widget builder which is only used when build get build in first time and
-  /// current state is [ServiceInitial]. When [onInitial] is omitted and
-  /// [buildWhen] is omitted, [fallback] widget will be used when first build
-  /// and layout would not update when state is [ServiceInitial]. That means
-  /// layout will be preserved from previous build.
-  final Widget Function(BuildContext context, ServiceInitial state)? onInitial;
-
-  /// A widget builder which is only used when current state is
-  /// [ServiceLoadInProgress]. When [onLoading] is omitted and [buildWhen] is
-  /// omitted, layout would not update when state is [ServiceLoadInProgress].
-  /// That means layout will be preserved from previous build.
-  final Widget Function(
-      BuildContext context,
-      ServiceLoadInProgress<ServiceRequestedEvent> state,
-      ServiceRequestedEvent event)? onLoading;
-
-  /// A widget builder which is only used when current state is
-  /// [ServiceLoadSuccess].
-  final Widget Function(
-      BuildContext context,
-      ServiceLoadSuccess<ServiceRequestedEvent, Response> state,
-      ServiceRequestedEvent event,
-      Response response) onSuccess;
-
-  /// A widget builder which is only used when current state is
-  /// [ServiceLoadFailure]. When [onFailure] is omitted and [buildWhen] is
-  /// omitted, layout would not update when state is [ServiceLoadFailure]. That
-  /// means layout will be preserved from previous build.
-  final Widget Function(
-      BuildContext context,
-      ServiceLoadFailure<ServiceRequestedEvent> state,
-      ServiceRequestedEvent event,
-      dynamic error)? onFailed;
-
-  /// A fallback widget for [builder] to use when [buildWhen] got passed and
-  /// widget builder function is omitted.
-  ///
-  /// Please make sure, which [RenderObject] is extends by parent widget. You
-  /// should the same type of [RenderObject] with [fallback].
-  /// For example: [RenderBox], [RenderSliver].
-  final Widget? fallback;
-
-  /// A constructor for creating a [ServiceBlocBuilder]. Default [buildWhen] is
-  /// used when [buildWhen] is omitted.
+  /// A constructor for creating a [ServiceBlocBuilder] with predefined state
+  /// widget builder.
   ServiceBlocBuilder({
     Key? key,
     Bloc? bloc,
     BlocBuilderCondition? buildWhen,
     this.onInitial,
     this.onLoading,
-    required this.onSuccess,
+    required this.onSucceed,
     this.onFailed,
-    this.fallback,
+    this.fallback = const SizedBox(),
   }) : super(
           key: key,
           bloc: bloc,
@@ -93,14 +51,14 @@ class ServiceBlocBuilder<
           builder: (context, state) {
             if (state is ServiceInitial) {
               if (onInitial == null) {
-                return fallback ?? Container();
+                return fallback;
               }
               return onInitial(context, state);
             }
 
             if (state is ServiceLoadInProgress<ServiceRequestedEvent>) {
               if (onLoading == null) {
-                return fallback ?? Container();
+                return fallback;
               }
               return onLoading(context, state, state.event);
             }
@@ -108,20 +66,68 @@ class ServiceBlocBuilder<
             if (state is ServiceLoadSuccess<ServiceRequestedEvent, Response>) {
               final data = state.data;
               if (data == null) {
-                return fallback ?? Container();
+                return fallback;
               }
 
-              return onSuccess(context, state, state.event, data);
+              return onSucceed(context, state, state.event, data);
             }
 
             if (state is ServiceLoadFailure<ServiceRequestedEvent>) {
               if (onFailed == null) {
-                return fallback ?? Container();
+                return fallback;
               }
               return onFailed(context, state, state.event, state.error);
             }
 
-            return fallback ?? Container();
+            return fallback;
           },
         );
+
+  /// A widget builder which is only called when build on first time or
+  /// [buildWhen] is omitted or custom [buildWhen] is passed and [onInitial] is
+  /// not omitted and current state is [ServiceInitial].
+  ///
+  /// Otherwise, [fallback] widget will be used when build on first time or
+  /// layout would not update when state is [ServiceInitial], which means layout
+  /// will be preserved from previous build.
+  final Widget Function(BuildContext context, ServiceInitial state)? onInitial;
+
+  /// A widget builder which is only called when [buildWhen] is omitted or
+  /// custom [buildWhen] is passed and [onLoading] is not omitted and current
+  /// state is [ServiceLoadInProgress].
+  ///
+  /// Otherwise, layout would not update when state is [ServiceInitial], which
+  /// means layout will be preserved from previous build.
+  final Widget Function(
+      BuildContext context,
+      ServiceLoadInProgress<ServiceRequestedEvent> state,
+      ServiceRequestedEvent event)? onLoading;
+
+  /// A widget builder which is only called when [buildWhen] is omitted or
+  /// custom [buildWhen] is passed and current state is [ServiceLoadSuccess].
+  final Widget Function(
+      BuildContext context,
+      ServiceLoadSuccess<ServiceRequestedEvent, Response> state,
+      ServiceRequestedEvent event,
+      Response response) onSucceed;
+
+  /// A widget builder which is only called when [buildWhen] is omitted or
+  /// custom [buildWhen] is passed [onFailed] is not omitted and current state
+  /// is [ServiceLoadFailure].
+  ///
+  /// Otherwise, layout would not update when state is [ServiceLoadFailure],
+  /// which means layout will be preserved from previous build.
+  final Widget Function(
+      BuildContext context,
+      ServiceLoadFailure<ServiceRequestedEvent> state,
+      ServiceRequestedEvent event,
+      dynamic error)? onFailed;
+
+  /// A fallback widget for [builder] to use when build on first time or
+  /// [buildWhen] got passed but widget builder function is omitted.
+  ///
+  /// Please make sure, which [RenderObject] is extends by parent widget. You
+  /// should the same type of [RenderObject] with [fallback].
+  /// For example: [RenderBox], [RenderSliver].
+  final Widget fallback;
 }