+++ service_bloc/example/lib/modal.dart
@@ -1,34 +1,8 @@
-import 'dart:async';
-import 'dart:convert';
-
-import 'package:bloc/bloc.dart';
 import 'package:built_value/built_value.dart';
 import 'package:built_value/serializer.dart';
 import 'package:example/serializer.dart';
-import 'package:http/http.dart';
-import 'package:service_bloc/service_bloc.dart';
 
-part 'example.g.dart';
-
-void main() async {
-  final organizationDetailServiceBloc = GithubOrganisationDetailServiceBloc();
-  organizationDetailServiceBloc
-      .add(GithubOrganisationDetailServiceRequested(name: 'flutter'));
-  await for (final state in organizationDetailServiceBloc.stream) {
-    if (state is! ServiceResponseState) continue;
-
-    if (state is ServiceLoadSuccess<GithubOrganisationDetailServiceRequested,
-        GithubOrganizationDetail>) {
-      print('response: ${state.data}');
-    }
-
-    if (state is ServiceLoadFailure) {
-      print('error: ${state.error}');
-    }
-
-    break;
-  }
-}
+part 'modal.g.dart';
 
 abstract class GithubOrganizationDetail
     implements
@@ -122,31 +96,4 @@ abstract class GithubOrganizationDetail
 
   static GithubOrganizationDetail fromJson(Map<String, dynamic> json) =>
       serializers.deserializeWith(serializer, json)!;
-}
-
-class GithubOrganisationDetailServiceRequested extends ServiceRequested {
-  const GithubOrganisationDetailServiceRequested({required this.name})
-      : assert(name.length > 0);
-
-  final String name;
-
-  @override
-  List<Object?> get props => [name];
-}
-
-class GithubOrganisationDetailServiceBloc extends ServiceBloc<
-    GithubOrganisationDetailServiceRequested, GithubOrganizationDetail> {
-  @override
-  FutureOr<void> onRequest(GithubOrganisationDetailServiceRequested event,
-      Emitter<ServiceState> emit) async {
-    try {
-      final response =
-          await get(Uri.parse('https://api.github.com/orgs/${event.name}'));
-      final organizationDetail =
-          GithubOrganizationDetail.fromJson(json.decode(response.body));
-      emit(ServiceLoadSuccess(event: event, data: organizationDetail));
-    } catch (error) {
-      emit(ServiceLoadFailure(event: event, error: error));
-    }
-  }
 }