+++ flutter_service_bloc/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.2
+
+* add documentation
+* add onInitial on service bloc listener
+* add default listenWhen on service bloc listener
+* add example application
+
 ## 0.0.1+1
 
 * add export service bloc library