+++ service_bloc/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.1+3
 
 * update README.md
+* update example
 
 ## 0.0.1+2